fix(point): 메서드명 수정 + repository 타입 수정
processPurchase -> deductPointForPurchase

+++ src/main/java/com/popcorntalk/domain/point/repository/PointRecordRepository.java
@@ -3,6 +3,6 @@ package com.popcorntalk.domain.point.repository;
 import com.popcorntalk.domain.point.entity.PointRecord;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PointRecordRepository extends JpaRepository<PointRecord, Integer> {
+public interface PointRecordRepository extends JpaRepository<PointRecord, Long> {
 
 }
